Move everything to GettingStarted Package
