+++ public/app/features/provisioning/GettingStarted/FeatureCard.tsx
