+++ public/app/features/provisioning/GettingStarted/FeatureList.tsx
