+++ public/app/features/provisioning/GettingStarted/SetupModal.tsx
