+++ public/app/features/provisioning/GettingStarted/SetupStep.tsx
