+++ public/app/features/provisioning/GettingStarted/Sidebar.tsx
