+++ public/app/features/provisioning/GettingStarted/types.ts
