+++ public/app/features/provisioning/GettingStarted/utils.ts
