+++ public/app/features/provisioning/constants.ts
@@ -1,3 +1,4 @@
 export const PROVISIONING_URL = '/admin/provisioning';
 export const CONNECT_URL = `${PROVISIONING_URL}/connect`;
 export const MIGRATE_URL = `${PROVISIONING_URL}/migrate`;
+export const GETTING_STARTED_URL = `${PROVISIONING_URL}/getting-started`;
